% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CellChat2Seurat}
\alias{CellChat2Seurat}
\title{Convert a CellChat object into a Seurat object}
\usage{
CellChat2Seurat(
  object,
  counts = NULL,
  group.by = NULL,
  assay = "Spatial",
  check.object = TRUE
)
}
\arguments{
\item{object}{CellChat object}

\item{counts}{Alternatively, use user-input `counts` for creating Seurat object instead of `object@data`}

\item{group.by}{Name of one metadata columns to group (color) cells. Default is the defined cell groups in CellChat object}

\item{assay}{Name of the initial assay. Default is "Spatial" when working on spatial imaging data}

\item{check.object}{whether checking the converted Seurat object. Set check.object = FALSE when not visualizing spatial clustering}
}
\value{
a Seurat object
}
\description{
Convert a CellChat object into a Seurat object
}
