% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cli.symbol}
\alias{cli.symbol}
\title{Generate 4 kinds of cli symbols widely used in CellChat}
\usage{
cli.symbol(symbol = NULL)
}
\arguments{
\item{symbol}{A string in `c('success','fail','info','output')` or a integer from `1:4` respectively}
}
\description{
Generate 4 kinds of cli symbols: 'success','fail','info','output'
}
\examples{
cli.symbol();cli.symbol(1);cli.symbol("info")
}
