% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{communicationDistPlot2}
\alias{communicationDistPlot2}
\title{Plot the cell-cell communication distance distribution}
\usage{
communicationDistPlot2(
  object,
  signaling.type = c("All", "Secreted", "Contact"),
  enriched.only = TRUE,
  density.alpha = 0.5
)
}
\arguments{
\item{object}{CellChat object}

\item{signaling.type}{the type of signaling}

\item{enriched.only}{whether to only show the communication distance distribution of the identified significant communication}

\item{density.alpha}{the transparence of the density plot}
}
\description{
Plot the cell-cell communication distance distribution
}
