% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeAvgCommunProb}
\alias{computeAvgCommunProb}
\title{Compute group-level cell-cell communication}
\usage{
computeAvgCommunProb(
  object,
  group.by = NULL,
  avg.type = c("avg", "sum"),
  min.percent = 0.1,
  min.cells.sr = 5,
  do.permutation = T,
  nboot = 100,
  seed.use = 1L,
  colocalization.use = F,
  thresh.colo = 0.05
)
}
\arguments{
\item{object}{SpatialCellChat object with communication probabilities for pairwise individual cells}

\item{group.by}{cell group information used for computing average communication probabilities}

\item{avg.type}{methods for integrating communication probabilities per cell group}

\item{min.percent}{Numeric from 0 to 1. Minimum percentage of expressed ligands or receptors per cell group to require for computing the group-level signaling
Default is 0.1.}

\item{min.cells.sr}{Integer greater than 0. Minimum number of cells required as senders or receivers per cell group for computing the group-level signaling
Default is 5.}

\item{do.permutation}{whether performing permutation test}

\item{nboot}{the number of permutations}

\item{seed.use}{set a random seed. By default, set the seed to 1.}

\item{colocalization.use}{whether filtering out spatially distant cell groups based on colocalization analysis between any cell groups}

\item{thresh.colo}{removal of cell-cell communication with no significant colocalizations (fdr < 0.05)}
}
\value{
A CellChat object with updated slot 'net':

object@net$prob is the inferred group-level communication probability (strength) array, where the first, second and third dimensions represent a source group, target group and ligand-receptor pair, respectively.

USER can access all the inferred cell-cell communications using the function 'subsetCommunication(object)', which returns a data frame.

object@net$pval is the corresponding p-values of each interaction
}
\description{
Compute group-level cell-cell communication
}
