% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeAvgCommunProb_LR_Avg}
\alias{computeAvgCommunProb_LR_Avg}
\title{Compute average communication probabilities of pairwise cell groups for one particular ligand-receptor pair/signaling pathway}
\usage{
computeAvgCommunProb_LR_Avg(
  prob,
  group,
  dataLR = NULL,
  min.percent = 0.1,
  min.cells.sr = 5
)
}
\arguments{
\item{prob}{a matrix of communication probabilities for pairwise individual cells for one particular ligand-receptor pair/signaling pathway}

\item{group}{Character vector. Cell group information used for computing averaged communication probabilities}

\item{dataLR}{a nCell*2 data matrix of a given pair of ligand-receptor.}

\item{min.percent}{Numeric from 0 to 1. Minimum percentage of expressed ligands or receptors per cell group to require for computing the group-level signaling
Default is 0.1.}

\item{min.cells.sr}{Integer greater than 0. Minimum number of cells required as senders or receivers per cell group for computing the group-level signaling
Default is 5.}
}
\value{
Returns a matrix containing the interaction weights between any two cell groups.
}
\description{
Compute average communication probabilities of pairwise cell groups for one particular ligand-receptor pair/signaling pathway
}
