% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeAvgCommunProb_LR_Visium_Avg}
\alias{computeAvgCommunProb_LR_Visium_Avg}
\title{computeAvgCommunProb_LR_Visium_Avg}
\usage{
computeAvgCommunProb_LR_Visium_Avg(
  prob,
  cell.type.proportion,
  cell.type.onehot
)
}
\arguments{
\item{prob}{a matrix of communication probabilities for pairwise individual cells for one particular ligand-receptor pair/signaling pathway}

\item{cell.type.proportion}{Ncells x NTypes Matrix, which comes from any kind of cell type decomposition method.}

\item{cell.type.onehot}{Ncells x NTypes Matrix, which differs from \code{cell.type.proportion}, can get through \code{cell.type.onehot <- 1*(cell.type.proportion>0)}}
}
\value{
A NTypes x NTypes matrix of communication probabilities for pairwise individual cell types for one particular ligand-receptor pair/signaling pathway
}
\description{
Compute average communication probabilities of pairwise cell groups for one particular ligand-receptor pair/signaling pathway
for 10X Visium Low Definition.
}
