% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeAvgCommunProb_LR_Visium_Sum}
\alias{computeAvgCommunProb_LR_Visium_Sum}
\title{computeAvgCommunProb_LR_Visium_Sum}
\usage{
computeAvgCommunProb_LR_Visium_Sum(
  prob,
  cell.type.proportion,
  cell.type.onehot = NULL
)
}
\arguments{
\item{prob}{a matrix of communication probabilities for pairwise individual cells for one particular ligand-receptor pair/signaling pathway}

\item{cell.type.proportion}{Ncells x NTypes Matrix, which comes from any kind of cell type decomposition method.}

\item{cell.type.onehot}{NULL. Placeholder.}
}
\value{
A NTypes x NTypes matrix of communication probabilities for pairwise individual cell types for one particular ligand-receptor pair/signaling pathway
}
\description{
Compute average communication probabilities of pairwise cell groups for one particular ligand-receptor pair/signaling pathway
for 10X Visium Low Definition.
}
