% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{computeCellDistance}
\alias{computeCellDistance}
\title{compute cell-cell distance and cell-cell contact adjacency matrix}
\usage{
computeCellDistance(
  coordinates,
  interaction.range = 250,
  contact.range = 10,
  ratio = NULL,
  tol = 10/2
)
}
\arguments{
\item{coordinates}{a data matrix in which each row gives the spatial locations/coordinates of each cell/spot}

\item{interaction.range}{Numeric(must positive). The maximum interaction/diffusion range of ligands. This hard threshold is used to filter out the connections between spatially distant cells}

\item{contact.range}{Numeric. The interaction range (Unit: microns) to restrict the contact-dependent signaling.
For spatial transcriptomics in a single-cell resolution, `contact.range` is approximately equal to the estimated cell diameter (i.e., the cell center-to-center distance), which means that contact-dependent and juxtacrine signaling can only happens when the two cells are contact to each other.
Typically, `contact.range = 10`, which is a typical human cell size. However, for low-resolution spatial data such as 10X visium, it should be the cell center-to-center distance (i.e., `contact.range = 100` for visium data).
Users can run the function `computeCellDistance` to get the center-to-center distance in the result's "d.spatial" key, which will help decide the value of `contact.range`.}

\item{ratio}{NULL or Numeric. ratio The conversion factor when converting spatial coordinates from Pixels or other units to Micrometers (i.e.,Microns).

For example, setting `ratio = 0.18` indicates that 1 pixel equals 0.18um in the coordinates.
For 10X visium, it is the ratio of the theoretical spot size (i.e., 65um) over the number of pixels that span the diameter of a theoretical spot size in the full-resolution image (i.e., 'spot.size.fullres' in the 'scalefactors_json.json' file).}

\item{tol}{Numeric. set a distance tolerance when computing cell-cell distances and contact adjacent matrix. Typically, `tol` should equal to the half value of cell/spot size in the unit of um.
For example, for 10X visium, `tol` can be set as `65/2`; for slide-seq, `tol` can be set as `10/2`.
If the cell/spot size is not known, we provide a function `computeCellDistance` to compute the center-to-center distance. `tol` can be the the half value of the minimum center-to-center distance.
By default `tol = 10/2`.}
}
\value{
List. A list has two keys: "d.spatial" and "adj.contact", storing cell-cell distance matrix and cell-cell contact adjacency matrix
}
\description{
Compute cell-cell distance based on the spatial coordinates and
generate cell-cell contact adjacency matrix with a contact.range restriction
}
