% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{computeColocalization}
\alias{computeColocalization}
\title{Assessment of the colocalization between any cell groups}
\usage{
computeColocalization(
  coordinates,
  group = NULL,
  idents.use = NULL,
  symmetric = TRUE,
  nboot = 100,
  seed.use = 1L
)
}
\arguments{
\item{coordinates}{a data matrix in which each row gives the spatial locations/coordinates of each cell/spot}

\item{group}{a factor vector defining the labels of each cell/spot}

\item{idents.use}{the identity to use}

\item{symmetric}{whether producing a symmetric colocalization matrix}

\item{nboot}{number of permutation when assessing the colocalization}

\item{seed.use}{set a random seed. By default, set the seed to 1.}
}
\value{
A square matrix giving the computed p-values after permuting cell labels
}
\description{
Assessment of the colocalization between any cell groups
}
