% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{computeCommunField}
\alias{computeCommunField}
\title{computeCommunField}
\usage{
computeCommunField(
  object,
  slot.name = "netP",
  signaling.name = NULL,
  top = 0.8,
  sparse = T
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{slot.name}{the slot name of object that is used for analysis}

\item{signaling.name}{alternative signaling pathway name used for analysis. If NULL, all the signaling pathways will be used}

\item{top}{the cutoff for selecting top signalings}

\item{sparse}{whether to apply a sparse strategy for selecting top signalings}
}
\value{
A SpatialCellChat object
}
\description{
computeCommunField
}
