% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeCommunProb}
\alias{computeCommunProb}
\title{computeCommunProb}
\usage{
computeCommunProb(
  object,
  LR.use = NULL,
  raw.use = TRUE,
  Kh = 0.5,
  n = 1,
  distance.use = TRUE,
  tol = NULL,
  interaction.range = 250,
  scale.distance = 0.01,
  use.AGAN = T,
  contact.dependent = TRUE,
  contact.range = 10,
  contact.dependent.forced = FALSE
)
}
\arguments{
\item{object}{CellChat object}

\item{LR.use}{A subset of ligand-receptor interactions used in inferring communication network}

\item{raw.use}{Whether use the raw data (i.e., `object@data.signaling`) or the projected data (i.e., `object@data.project`).
Set raw.use = FALSE to use the projected data when analyzing single-cell data with shallow sequencing depth because the projected data could help to reduce the dropout effects of signaling genes, in particular for possible zero expression of subunits of ligands/receptors.}

\item{Kh}{Parameter in Hill function}

\item{n}{Parameter in Hill function}

\item{distance.use}{Whether to use distance constraints to compute communication probability. Setting `distance.use = TRUE` indicates that the cell-cell communication probability is inversely proportional to the computed distance.}

\item{tol}{NULL or Numeric. set a distance tolerance when computing cell-cell distances and contact adjacent matrix. By default `tol = NULL` means distance tolerance equals to spot.size/2 or cell.diameter/2.}

\item{interaction.range}{The maximum interaction/diffusion length of ligands (Unit: microns). This hard threshold is used to filter out the connections between spatially distant individual cells}

\item{scale.distance}{A scale or normalization factor for the spatial distances when setting `distance.use = TRUE`. For example, scale.distance equals 1, 0.1, 0.01, 0.001, 0.11, or 0.011. We choose this values such that the minimum value of the scaled distances is in [1,2]. This value is not necessary when setting `distance.use = FALSE`.}

\item{use.AGAN}{Boolean. Whether to take agonist (AG) and antagonist (AN) into consideration when calculating the intercellular communication probability}

\item{contact.dependent}{Boolean. Whether determining spatially proximal cell groups based on the `contact.range`. By default `contact.dependent = TRUE` when inferring contact-dependent and juxtacrine signaling (including ECM-Receptor and Cell-Cell Contact signaling classified in CellChatDB$interaction$annotation).
If only focusing on `Secreted Signaling`, the `contact.dependent` will be automatically set as FALSE except for `contact.dependent.forced = TRUE`.}

\item{contact.range}{Numeric. The interaction range (Unit: microns) to restrict the contact-dependent signaling.
For spatial transcriptomics in a single-cell resolution, `contact.range` is approximately equal to the estimated cell diameter (i.e., the cell center-to-center distance), which means that contact-dependent and juxtacrine signaling can only happens when the two cells are contact to each other.
Typically, `contact.range = 10`, which is a typical human cell size. However, for low-resolution spatial data such as 10X visium, it should be the cell center-to-center distance (i.e., `contact.range = 100` for visium data).  The function \link{computeCellDistance} can compute the center-to-center distance.}

\item{contact.dependent.forced}{Boolean. Whether forcing to determine spatially proximal cell regions based on the `contact.range` for all cells/spots in the CellChat Object.
Users can set `contact.dependent.forced = TRUE` to turn `Secreted Signaling` interactions into a contact manner.}
}
\value{
CellChat object
}
\description{
Compute the communication probability/strength between any interacting individual cells
}
