% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{computeGridSize}
\alias{computeGridSize}
\title{computeGridSize}
\usage{
computeGridSize(
  object,
  cellsize = NULL,
  grid.resolution = NULL,
  do.plot = T,
  what = "polygons",
  square = T
)
}
\arguments{
\item{object}{CellChat object}

\item{cellsize}{NULL or Numeric. If NULL, the function will tell you the default `cellsize` in the SpatialCellChat object.
If not NULL and be Numeric, the new cellsize will be equal to cellsize*grid.resolution.
Note: We use [sf::st_make_grid()] to make grid data, so the param is a numeric vector
of length 1 or 2 with target cellsize, please see details in [sf::st_make_grid()]:
for square or rectangular cells the width and height, for hexagonal cells the distance between
opposite edges (edge length is cellsize/sqrt(3)). A length units object can be passed,
or an area unit object with area size of the square or hexagonal cell.}

\item{grid.resolution}{Numeric. By default, it is set to be 2.}

\item{do.plot}{Boolean. If FALSE, only print hints in terminal.}

\item{what}{Character. One of: "polygons", "corners", or "centers". See details in [sf::st_make_grid()]}

\item{square}{Boolean. If FALSE, create hexagonal grid. See details in [sf::st_make_grid()]}
}
\value{
ggplot
}
\description{
Use this function to view SpatialCellChat object after the "grid" operation and visualize it.
}
