% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{createCellCellContactMatrixFrom_dspatial}
\alias{createCellCellContactMatrixFrom_dspatial}
\title{create a cell-cell contact matrix from `d.spatial` object}
\usage{
createCellCellContactMatrixFrom_dspatial(d.spatial, contact.threshold, tol)
}
\arguments{
\item{d.spatial}{a CsparseMatrix (dsC Matrix) object, see in `computeCellDistance`:
d.spatial <- stats::dist(coordinates);
d.spatial <- createSparseMatrixFrom_distObj(d.spatial)}

\item{contact.threshold}{Numeric. The interaction range threshold(Unit: microns) to restrict the contact-dependent signaling,
passed from `computeCellDistance`'s parameter `contact.range`.
If the distance of a pair of cells is larger than contact.threshold, regard the two cells as not in contact. Otherwise, regard the two cells as in contact.}

\item{tol}{Numeric. set a distance tolerance when computing cell-cell distances and contact adjacent matrix,
passed from `computeCellDistance`'s same parameter.
By default `tol = NULL` means `tol` equals to the half value of cell/spot size in the unit of um.}
}
\value{
A `CsparseMatrix` object in `Matrix` Package
}
\description{
A function defined for `computeCellDistance`. We use the function `createSparseMatrixFrom_distObj` to convert the dist object into a CsparseMatrix object when running `computeCellDistance`.
Here we design the function to generate cell-cell contact adjacency matrix with a `contact.range` restriction and a `tol` tolerance passed from `computeCellDistance`'s parameters respectively
}
