% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{createPspatialFrom_dspatial}
\alias{createPspatialFrom_dspatial}
\title{create P.spatial from d.spatial object}
\usage{
createPspatialFrom_dspatial(d.spatial, distance.use = T)
}
\arguments{
\item{d.spatial}{CsparseMatrix (dsC Matrix) object. A cell-cell distance matrix which has been scaled, see in \link{computeCommunProb}}

\item{distance.use}{Boolean. Whether to use distance constraints to compute communication probability, passed from \link{computeCommunProb}'s same parameter
Setting `distance.use = TRUE` indicates that the cell-cell communication probability is inversely proportional to the computed distance.
Otherwise, the cell-cell communication probability share the same computation factor 1 within the maximum interaction/diffusion length of ligands}
}
\value{
CsparseMatrix object in `Matrix` Package
}
\description{
A function defined for \link{computeCommunProb}. To model ligands' diffusion attenuation, we take the reciprocal of cell-cell distance
as a factor (named as P.spatial) determining the cell-cell communication probability. Here we design
the function to help take reciprocal of a `d.spatial` matrix.
}
