% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.H_to_conn_mat}
\alias{.H_to_conn_mat}
\title{Make sure H is cell x factor
Experimental method to examine the dispersion of results}
\usage{
.H_to_conn_mat(H)
}
\arguments{
\item{H}{a list of numeric matrices or a single matrix}
}
\description{
Make sure H is cell x factor
Experimental method to examine the dispersion of results
}
