% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.plotSuggestK}
\alias{.plotSuggestK}
\title{This function visualizes how objective error varies with different values of k.}
\usage{
.plotSuggestK(stats)
}
\arguments{
\item{stats}{a data frame containing evaluation results across different values of k}
}
\value{
a ggplot2 object
}
\description{
This function visualizes how objective error varies with different values of k.
}
