% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{elementwiseProductForSparseMatandDenseMat}
\alias{elementwiseProductForSparseMatandDenseMat}
\title{element-wise product for a sparse matrix and a dense matrix}
\usage{
elementwiseProductForSparseMatandDenseMat(SparseMat, DenseMat)
}
\arguments{
\item{SparseMat}{A dgCMatrix object in `Matrix` Package.}

\item{DenseMat}{A dense matrix shares the same dimensions with `SparseMat`}
}
\value{
element-wise product result. A dgCMatrix object in `Matrix` Package
}
\description{
A function defined for \link{computeCommunProb}. Define an element-wise product operation for a sparse matrix and a dense matrix,
to speed up the product operation
}
