% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{==.sparse3Darray}
\alias{==.sparse3Darray}
\title{Fast `==` operation for `sparse3Darray` Class}
\usage{
\method{==}{sparse3Darray}(arr1, arr2)
}
\arguments{
\item{arr1}{A sparse3Darray}

\item{arr2}{A sparse3Darray}
}
\value{
Boolean.
}
\description{
Fast `==` operation for `sparse3Darray` Class
}
