% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{filterCommunication}
\alias{filterCommunication}
\title{Filter cell-cell communication if there are only few number of cells in certain cell groups or only few interactions}
\usage{
filterCommunication(object, min.cells = 10, min.links = 5, min.cells.sr = 5)
}
\arguments{
\item{object}{CellChat object}

\item{min.cells}{the minimum number of cells required in each cell group for filtering cell group-level communication}

\item{min.links}{the minimum number of links/interactions required in the ligand-receptor pair for filtering individual cell-level communication}

\item{min.cells.sr}{the minimum number of cells required as senders or receivers for filtering individual cell-level communication}
}
\value{
CellChat object with an updated slot net
}
\description{
Filter cell-cell communication if there are only few number of cells in certain cell groups or only few interactions
}
