% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{filterProbability}
\alias{filterProbability}
\title{filterProbability}
\usage{
filterProbability(object, nboot = 100, seed.use = 666L, thresh = 0.05)
}
\arguments{
\item{object}{CellChat object}

\item{nboot}{Numeric. The number of bootstrap samples, 100 by default.}

\item{seed.use}{Integer. The random seed used when taking a sample from the communication probabilities}

\item{thresh}{Numeric. The threshold for defining significant individual cell-cell communication at (1-thresh) of a shuffled distribution of each L-R pair.}
}
\value{
CellChat object
}
\description{
Filter out statistically non-significant communication probability at the level of individual cells after running \link{computeCommunProb}
}
