% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_custom_colorscale}
\alias{generate_custom_colorscale}
\title{generate_custom_colorscale}
\usage{
generate_custom_colorscale(
  color.heatmap,
  n.colors = 8,
  reversescale = F,
  plotly.format = T
)
}
\arguments{
\item{color.heatmap}{a character string or vector indicating the colormap option to use. It can be the avaibale color palette in brewer.pal() or viridis_pal() (e.g., "Spectral","viridis")}

\item{n.colors}{number of basic colors to generate from color palette}

\item{reversescale}{whether to reverse the direction of the color scale}

\item{plotly.format}{whether to convert the colormap into a plotly-compatible format}
}
\value{
List. See the parameter `scene` in \code{plotly::layout}
}
\description{
generate_custom_colorscale
}
