% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_RedsBlues_colormap}
\alias{generate_RedsBlues_colormap}
\title{generate RedsBlues colormap for a numeric vector}
\usage{
generate_RedsBlues_colormap(vec, mid.color = "#E5E5E5", switch.RedsBlues = F)
}
\arguments{
\item{vec}{A numeric vector. Positive values will be drawn red,
negative values will be drawn blue.}

\item{mid.color}{A string representing the color mapping to the `0` in the numeric vector.}

\item{switch.RedsBlues}{Whether to reverse the colormap}
}
\value{
Function. Generate RedsBlues colors.
}
\description{
generate RedsBlues colormap for a numeric vector
}
