% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_dens}
\alias{get_dens}
\title{get_dens}
\usage{
get_dens(data, dens, method)
}
\arguments{
\item{data}{matrix with dimensions where to calculate the density from. Only
the first two dimensions will be used}

\item{dens}{a list with the density estimates from the selected method}

\item{method}{Kernel density estimation method:
\itemize{
\item \code{ks}: Computes density using the \code{kde} function from the
 \code{ks} package.
\item \code{wkde}: Computes density using a modified version of the
 \code{kde2d} function from the \code{MASS}
package to allow weights. Bandwidth selection from the \code{ks} package
 is used instead.
}}
}
\value{
a vector with corresponding densities for each observation
}
\description{
get_dens
}
