% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{HillFunctionFordataLR}
\alias{HillFunctionFordataLR}
\title{HillFunction for dataLR object}
\usage{
HillFunctionFordataLR(dataLR, Kh, n)
}
\arguments{
\item{dataLR}{A dgCMatrix object in `Matrix` Package.}

\item{Kh}{A hyper parameter in hill function, passed from \link{computeCommunProb}}

\item{n}{A hyper parameter in hill function, passed from \link{computeCommunProb}}
}
\value{
a normalized dataLR
}
\description{
A function defined for \link{computeCommunProb}. A Hill function was used to model the interactions
between L and R with an EC50 parameter Kh whose default value was set to be 0.5 as the input data has a normalized range from 0 to 1.
Previous hill function is `dataLR^n/(Kh^n + dataLR^n)`, notice dataLR is a sparse matrix, so we design the function to do element-wise hill function upon each non-zero value in the `dataLR` matrix
}
