% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{identifyCellTopics}
\alias{identifyCellTopics}
\title{identifyCellTopics}
\usage{
identifyCellTopics(
  object,
  slot.name = "net",
  pattern = c("incoming", "outgoing"),
  do.scale = F,
  topic.k = seq(5, 50, 5),
  n.reps = 10,
  prop = 0.33,
  tol = 1e-06,
  maxit = 600L,
  L1 = c(0, 0),
  verbose = T,
  seed.use = 666L
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{slot.name}{the slot name of object to be analyzed}

\item{pattern}{"outgoing" or "incoming"}

\item{do.scale}{whether to scale the matrix}

\item{topic.k}{integer vector. Choose several ranks for cross validation}

\item{n.reps}{integer. How many replications to use for each k in cross validation}

\item{prop}{numeric. Proportion(0<p<1) of the Mat's elements used for cross validation}

\item{tol}{numeric. Stopping criteria.}

\item{maxit}{integer. Stopping criteria.}

\item{L1}{L1/LASSO penalties between 0 and 1, array of length two for c(w, h)}

\item{verbose}{whether to print model tolerances between iterations}

\item{seed.use}{seed}
}
\value{
SpatialCellChat object, in which the specified slot is augmented with identified cell topics results (cell-by-topic and signaling-by-topic matrices)
or with intermediate evaluation results used for selecting the rank k
}
\description{
identifyCellTopics
}
