% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{identifyOverExpressedGenes}
\alias{identifyOverExpressedGenes}
\title{Identify over-expressed signaling genes associated with each cell group or spatially variable features independent of cell groups}
\usage{
identifyOverExpressedGenes(
  object,
  do.grid = F,
  cellsize = c(5, 5),
  what = "polygons",
  square = T,
  data.use = NULL,
  selection.method = c("wilcox", "moransi", "meringue"),
  features.name = "features",
  group.by = NULL,
  idents.use = NULL,
  invert = FALSE,
  group.dataset = NULL,
  pos.dataset = NULL,
  only.pos = TRUE,
  features = NULL,
  return.object = TRUE,
  thresh.pc = 0,
  thresh.fc = 0,
  thresh.p = 0.05
)
}
\arguments{
\item{object}{CellChat object}

\item{do.grid}{Boolean. Whether to do "grid" operation to speed up computation.}

\item{cellsize}{Numeric vector of length 1 or 2. We use [sf::st_make_grid()] to make grid data. It is a numeric vector
of length 1 or 2 with target cellsize, please see details in [sf::st_make_grid()]:
for square or rectangular cells the width and height, for hexagonal cells the distance between
opposite edges (edge length is cellsize/sqrt(3)). A length units object can be passed,
or an area unit object with area size of the square or hexagonal cell.

Tips: Please run `computeGridSize` to get a proper `cellsize`.the unit of the `cellsize` you input is "pixel" or a unit used in old CellChat object's coordinates}

\item{what}{Character. One of: "polygons", "corners", or "centers". See details in [sf::st_make_grid()]}

\item{square}{Boolean. If FALSE, create hexagonal grid. See details in [sf::st_make_grid()]}

\item{data.use}{a customed data matrix. Default: data.use = NULL and the expression matrix in the slot 'data.signaling' is used}

\item{selection.method}{Method for selecting (spatially) variable features.}

\item{features.name}{a char name used for storing the over-expressed signaling genes in `object@var.features[[features.name]]`}

\item{group.by}{cell group information; default is `object@idents`; otherwise it should be one of the column names of the meta slot}

\item{idents.use}{a subset of cell groups used for analysis}

\item{invert}{whether invert the idents.use}

\item{group.dataset}{dataset origin information in a merged CellChat object; set it as one of the column names of meta slot when identifying the highly enriched genes in one dataset for each cell group}

\item{pos.dataset}{the dataset name used for identifying highly enriched genes in this dataset for each cell group}

\item{only.pos}{Only return positive markers}

\item{features}{features used for identifying Over Expressed genes. default use all features}

\item{return.object}{whether return the object; otherwise return a data frame consisting of over-expressed signaling genes associated with each cell group}

\item{thresh.pc}{Threshold of the percent of cells expressed in one cluster}

\item{thresh.fc}{Threshold of Log Fold Change}

\item{thresh.p}{Threshold of p-values}
}
\value{
A CellChat object or a data frame. If returning a CellChat object, two new elements named 'features.name' and paste0(features.name, ".info") will be added into the list `object@var.features`
`object@var.features[[features.name]]` is a vector consisting of the identified over-expressed signaling genes;
`object@var.features[[paste0(features.name, ".info")]]` is a data frame returned from the differential expression analysis / the simple filtering
}
\description{
USERS can use customized gene set as over-expressed signaling genes by setting `object@var.features[[features.name]] <- features.sig`
The Bonferroni corrected/adjusted p value can be obtained via `object@var.features[[paste0(features.name, ".info")]]`. Note that by default `features.name = "features"`
}
