% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{identifyTopicCommunities}
\alias{identifyTopicCommunities}
\title{Identifying topic communities}
\usage{
identifyTopicCommunities(
  object,
  slot.name = "net",
  pattern = c("incoming", "outgoing"),
  resolution = NULL,
  clustering.method = c("seurat", "autoStardust", "weightStardust", "fusion"),
  nPrograms = NULL,
  k.signaling = 20,
  k.spatial = 20,
  method = "matrix",
  algorithm = 1L,
  prune.SNN = 1/15,
  graph.name = NULL,
  spatialWeight = 0.5,
  k.fusion = 20,
  t.fusion = 20,
  truncated = FALSE,
  resRange = NULL,
  eigengap = TRUE,
  tol = 0.01,
  min.pct = 0.1,
  logfc.threshold = 0.25,
  do.plot = TRUE,
  filename = "",
  seed.use = 666L,
  return.seurat = T
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{slot.name}{the slot name of object to be analyzed}

\item{pattern}{"outgoing" or "incoming"}

\item{resolution}{the resolution in Leiden algorithm; if it is NULL, the optimal resoultion will be inferred based on eigen spectrum}

\item{clustering.method}{method for performing subclustering using the combination of spatial location and gene expression information ("fusion") or the gene expression only ("seurat")}

\item{nPrograms}{the number of PC dimensions to use for computing SNN based on the gene expression}

\item{k.signaling}{the number of nearest neighbors when computing SNN based on the gene expression}

\item{k.spatial}{the number of nearest neighbors when computing SNN based on the spatial location}

\item{method}{method for running leiden (defaults to matrix which is fast for small datasets). Enable method = "igraph" to avoid casting large data to a dense matrix.}

\item{algorithm}{algorithm for modularity optimization (1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm). Leiden requires the leidenalg python.}

\item{prune.SNN}{sets the cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Any edges with values less than or equal to this will be set to 0 and removed from the SNN graph. Essentially sets the strigency of pruning (0 — no pruning, 1 — prune everything).}

\item{graph.name}{name of graph to use for the clustering algorithm}

\item{spatialWeight}{spot distance weight when running weightStardust clustering method. It is in [0,1]. Weight for the linear transformation of spot distance
1 means spot distance weight as much as profile distance, 0 means spot distance doesn't contribute at all in the overall
distance measure}

\item{k.fusion}{number of neighbors in K-nearest neighbors part of the fusion algorithm}

\item{t.fusion}{number of iterations for the diffusion process in the fusion algorithm}

\item{truncated}{whether truncating the fusion similarity matrix by only retaining the k-largest values when using the "fusion" clustering method}

\item{resRange}{the range of resolutions in Leiden algorithm; if it is NULL, the optimal range of resoultion will be from 0.02 to 1.5}

\item{eigengap}{whether determining the number of subclusters using the eigengap method}

\item{tol}{the tolerance value when determining the number of subclusters}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations}

\item{logfc.threshold}{limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells}

\item{do.plot}{whether to generate and save visualization results for identified topic communities}

\item{filename}{a character string used as a suffix for the names of output image files if `do.plot=TRUE`}

\item{seed.use}{random seed}

\item{return.seurat}{whether to store the generated Seurat object in the input object}
}
\value{
SpatialCellChat object
}
\description{
Identifying topic communities
}
