% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCellChat_class.R
\name{level.SpatialCellChat}
\alias{level.SpatialCellChat}
\title{getIdent}
\usage{
level.SpatialCellChat(object)
}
\arguments{
\item{object}{A SpatialCellChat object}
}
\value{
Character string. SpatialCellChat's idents saved in the `idents` slot.
}
\description{
getIdent
}
