% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCellChat_class.R
\name{liftSpatialCellChat}
\alias{liftSpatialCellChat}
\title{Update a SpatialCellChat object by lifting up the cell groups to the same cell labels across all datasets}
\usage{
liftSpatialCellChat(object, group.new = NULL)
}
\arguments{
\item{object}{A single or merged SpatialCellChat object}

\item{group.new}{A char vector giving the cell labels to lift up. The order of cell labels in the vector will be used for setting the new cell identity.

 If the input is a merged SpatialCellChat object and group.new = NULL, it will use the cell labels from one dataset with the maximum number of cell groups

 If the input is a single SpatialCellChat object, `group.new` must be defined.}
}
\value{
a updated SpatialCellChat object
}
\description{
This function is useful when comparing inferred communications across different datasets with different cellular compositions
}
