% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{makeGridSpatialCellChat}
\alias{makeGridSpatialCellChat}
\title{makeGridSpatialCellChat}
\usage{
makeGridSpatialCellChat(
  object,
  data.slot = c("data", "data.raw", "data.signaling", "data.scale", "data.project"),
  cellsize = c(5, 5),
  what = "polygons",
  square = T,
  idents.ties.method = "first"
)
}
\arguments{
\item{object}{CellChat object}

\item{data.slot}{A CellChat object may have 5 slots to store data matrix:
data.raw,data,data.signaling,data.scale,data.project, choose one data slot
to make a new grid CellChat object. By default, use "data".}

\item{cellsize}{Numeric vector of length 1 or 2. We use [sf::st_make_grid()] to make grid data. It is a numeric vector
of length 1 or 2 with target cellsize, please see details in [sf::st_make_grid()]:
for square or rectangular cells the width and height, for hexagonal cells the distance between
opposite edges (edge length is cellsize/sqrt(3)). A length units object can be passed,
or an area unit object with area size of the square or hexagonal cell.

Tips: Please run `computeGridSize` to get a proper `cellsize`.the unit of the `cellsize` you input is "pixel" or a unit used in old CellChat object's coordinates}

\item{what}{Character. One of: "polygons", "corners", or "centers". See details in [sf::st_make_grid()]}

\item{square}{Boolean. If FALSE, create hexagonal grid. See details in [sf::st_make_grid()]}

\item{idents.ties.method}{To determine the ident/cell group of each grid, we use [base::max.col()]
to choose a ident containing the most cells per grid. See `ties.method` in [base::max.col()].}
}
\value{
CellChat object
}
\description{
makeGridSpatialCellChat
}
