% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mergeChatIntoSeurat}
\alias{mergeChatIntoSeurat}
\title{Merge SpatialCellChat objects}
\usage{
mergeChatIntoSeurat(
  object.list,
  slot.name = "net",
  pattern = c("incoming", "outgoing"),
  normalize.chat = T,
  normalization.method = "LogNormalize",
  merge.chat = F
)
}
\arguments{
\item{object.list}{A list of multiple SpatialCellChat objects.}

\item{slot.name}{Which slot in SpatialCellChat used to merge.}

\item{pattern}{SpatialCellChat pattern used to merge. In detail, it comes from
`centr.cell` in net or netP slot.
If the `object.list` is a SpatialCellChat object, the functions will return a seurat
object combining both "indeg" and "outdeg" in `centr.cell`.
If the `object.list` is a SpatialCellChat object list, the functions will return a seurat
object combining "indeg" or "outdeg" in `centr.cell` from objects respectively.}

\item{normalize.chat}{Whether to normalize SpatialCellChat pattern with \code{\link[Seurat]{NormalizeData.Seurat}}.}

\item{normalization.method}{Inherit from \code{\link[Seurat]{NormalizeData.Seurat}}.}

\item{merge.chat}{Use it when the `object.list` is a SpatialCellChat object, whether to
merge both "indeg" and "outdeg" in `centr.cell` into only one assay data in seurat.}
}
\value{
SeuratObject.
}
\description{
Merge SpatialCellChat objects
}
