% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCellChat_class.R
\name{mergeSpatialCellChat}
\alias{mergeSpatialCellChat}
\title{Merge SpatialCellChat objects}
\usage{
mergeSpatialCellChat(
  object.list,
  add.names = NULL,
  merge.data = FALSE,
  cell.prefix = FALSE,
  show.plot = T
)
}
\arguments{
\item{object.list}{A list of multiple SpatialCellChat objects.}

\item{add.names}{A vector containing the name of each dataset.}

\item{merge.data}{Whether to merge the `data` slot for ALL genes. Default only merges the data of signaling genes.}

\item{cell.prefix}{Whether to prefix cell names.}

\item{show.plot}{Whether to show merged plots.}
}
\description{
Merge SpatialCellChat objects
}
