% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCellChat_class.R
\name{misc}
\alias{misc}
\alias{misc<-}
\alias{misc.SpatialCellChat}
\alias{misc<-.SpatialCellChat}
\title{Get and set miscellaneous data}
\usage{
misc(object, ...)

misc(object, ...) <- value

\method{misc}{SpatialCellChat}(object, key = NULL, ...)

\method{misc}{SpatialCellChat}(object, key = NULL, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Data to add}

\item{key}{Name of specific bit of meta data to pull}
}
\value{
miscellaneous data

An object with miscellaneous data added
}
\description{
Get and set miscellaneous data
}
\concept{data-access}
