% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{modifySparse3Darray}
\alias{modifySparse3Darray}
\title{modifySparse3Darray}
\usage{
modifySparse3Darray(array3d, do.binary = T, cutoff = NULL, remain.cutoff.v = T)
}
\arguments{
\item{array3d}{sparse3Darray. See details in \code{\link[spatstat.sparse]{sparse3Darray}}}

\item{do.binary}{Boolean. Use 1 to replace the non-zero values in the 3Darray.}

\item{cutoff}{Numeric. Cut off the sparse3Darray's values to make
array' values >= cutoff.}

\item{remain.cutoff.v}{Boolean. Use `>=` or `>`.}
}
\value{
sparse3Darray.
}
\description{
modifySparse3Darray
}
