% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_as_sparse3Darray}
\alias{my_as_sparse3Darray}
\title{my_as_sparse3Darray}
\usage{
my_as_sparse3Darray(x, ..., strict = FALSE, nonzero = FALSE)
}
\arguments{
\item{x}{Data in another format (see Details in \code{\link[spatstat.sparse]{as.sparse3Darray}}.}

\item{...}{Ignored.}

\item{strict}{Boolean. Logical value specifying whether to enforce the rule that each entry in i,j,k,x refers to a different cell. If strict=TRUE, entries which refer to the same cell in the array will be reduced to a single entry by summing the x values. Default is strict=FALSE.}

\item{nonzero}{Boolean. Logical value specifying whether to remove any entries of x which equal zero.}
}
\value{
Sparse three-dimensional array (object of class "sparse3Darray").
}
\description{
A faster realization in comparison to \code{\link[spatstat.sparse]{as.sparse3Darray}}
}
