% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_future_lapply}
\alias{my_future_lapply}
\title{my_future_lapply}
\usage{
my_future_lapply(
  X,
  FUN,
  ...,
  future.seed = T,
  simplify = F,
  hint.message = "Computing...",
  strategy.message = TRUE,
  future.label = "future_lapply-\%d"
)
}
\arguments{
\item{X}{A vector-like object to iterate over.}

\item{FUN}{A function taking at least one argument.}

\item{...}{(optional) Additional arguments passed to `FUN()`.
For `future_*apply()` functions and `replicate()`, any `future.*` arguments
part of `\ldots` are passed on to `future_lapply()` used internally.}

\item{future.seed}{A logical or an integer (of length one or seven),
or a list of `length(X)` with pre-generated random seeds.
For details, see \code{\link[future.apply]{future_lapply}}.}

\item{simplify}{different from the `simplify` parameter in `my_future_sapply`.
If simplify = TRUE, we run \code{\link[base]{unlist}} upon the result.}

\item{hint.message}{A hint message shown after the progress bar. By default,
the message is "Computing..."}

\item{strategy.message}{Whether to show the future stretegy information}

\item{future.label}{If a character string, then each future is assigned
a label `sprintf(future.label, chunk_idx)`.  If TRUE, then the
same as `future.label = "future_lapply-%d"`.  If FALSE, no labels
are assigned.}
}
\value{
For `future_lapply()`, a list with same length and names as `X`.
See \code{\link[base]{lapply}} for reference.
}
\description{
future_lapply with progress bar and stable random number generation
}
\examples{
my_future_lapply(X=1:10,FUN = function(x){Sys.sleep(11-x);sqrt(x)})
}
