% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_future_sapply}
\alias{my_future_sapply}
\title{my_future_sapply}
\usage{
my_future_sapply(
  X,
  FUN,
  ...,
  simplify = TRUE,
  USE.NAMES = TRUE,
  future.envir = parent.frame(),
  future.seed = T,
  future.label = "future_sapply-\%d",
  hint.message = "Computing...",
  strategy.message = TRUE
)
}
\arguments{
\item{X}{A vector-like object to iterate over.}

\item{FUN}{A function taking at least one argument.}

\item{...}{(optional) Additional arguments passed to `FUN()`.
For `future_*apply()` functions and `replicate()`, any `future.*` arguments
part of `\ldots` are passed on to `future_lapply()` used internally.}

\item{simplify}{See \code{\link[base]{sapply}}, respectively.}

\item{USE.NAMES}{See \code{\link[base]{sapply}}.}

\item{future.envir}{An [environment] passed as argument `envir` to
\code{\link[future]{future}} as-is.}

\item{future.seed}{A logical or an integer (of length one or seven),
or a list of `length(X)` with pre-generated random seeds.
For details, see \code{\link[future.apply]{future_lapply}}.}

\item{future.label}{If a character string, then each future is assigned
a label `sprintf(future.label, chunk_idx)`.  If TRUE, then the
same as `future.label = "future_sapply-%d"`.  If FALSE, no labels
are assigned.}

\item{hint.message}{A hint message shown after the progress bar. By default,
the message is "Computing..."}

\item{strategy.message}{Whether to show the future stretegy information}
}
\value{
For `future_sapply()`, a vector with same length and names as \code{X}.
See \code{\link[base]{sapply}} for reference.
}
\description{
future_sapply with progress bar and stable random number generation
}
\examples{
my_future_sapply(X=1:10,FUN = function(x){Sys.sleep(11-x);sqrt(x)})
}
