% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_obj_size}
\alias{my_obj_size}
\title{my_obj_size}
\usage{
my_obj_size(...)
}
\arguments{
\item{...}{Any R object. See details in \code{\link[pryr]{object_size}}.}
}
\value{
String. Object size easy to read.
}
\description{
works similarly to object.size, but counts more accurately and includes the size of environments.
}
\examples{
\dontrun{
my_obj_size(list(1,2,3))
}
}
