% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{myElementwiseProduct}
\alias{myElementwiseProduct}
\title{element-wise product for a sparse matrix and a dense vector's crossprod}
\usage{
myElementwiseProduct(SparseMat, DenseVec)
}
\arguments{
\item{SparseMat}{A dgCMatrix object in `Matrix` Package.}

\item{DenseVec}{A numeric vector shares the same dimensions with `SparseMat`}
}
\value{
element-wise product result. A dgCMatrix object in `Matrix` Package
}
\description{
A function defined for \link{computeCommunProb}. Define an element-wise product operation for a sparse matrix and a dense vector's crossprod,
to speed up the product operation
}
