% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{netAnalysis_computeCentrality}
\alias{netAnalysis_computeCentrality}
\title{Compute the network centrality scores allowing identification of dominant senders, receivers, mediators and influencers in all inferred communication networks}
\usage{
netAnalysis_computeCentrality(
  object = NULL,
  net = NULL,
  slot.name = "net",
  signaling.name = NULL,
  do.group = F,
  thresh = 0.05,
  degree.only = T
)
}
\arguments{
\item{object}{CellChat object; If object = NULL, USER must provide `net`}

\item{net}{compute the centrality measures on a specific signaling network given by a 2 or 3 dimemsional array net}

\item{slot.name}{the slot name of object that is used to compute centrality measures of signaling networks}

\item{signaling.name}{provide signalings name (LR pairs or pathways) to subset the communication net}

\item{do.group}{set `do.group = TRUE` when computing centrality of each cell group; set `do.group = FALSE` when computing centrality of each individual cell}

\item{thresh}{threshold of the p-value for determining significant interaction}

\item{degree.only}{To speed up computation, run `netAnalysis_computeCentrality` with `degree.only = T` when user only needs
"outdeg_unweighted", "indeg_unweighted","outdeg","indeg","page_rank" to do analysis}
}
\description{
NB: This function was previously named as `netAnalysis_signalingRole`.  The previous function `netVisual_signalingRole` is now named as `netAnalysis_signalingRole_network`.
}
