% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual_CommunField}
\alias{netVisual_CommunField}
\title{netVisual_CommunField}
\usage{
netVisual_CommunField(
  object,
  signaling,
  slot.name = "netP",
  pattern = c("incoming", "outgoing"),
  zoom.to = 0L,
  zoom.mode = T,
  anchor.x = 0.5,
  anchor.y = 0.5,
  color.use = NULL,
  point.size = 2,
  min.mag = 0,
  alpha.image = 0.2,
  arrow.angle = 15,
  arrow.length = 0.6,
  arrow.size = 1,
  arrow.color = NULL,
  ...,
  legend.size = 0.5,
  legend.text.size = 8,
  plot.title.name = NULL,
  plot.title.text.size = 8
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{signaling}{a signaling pathway or ligand-receptor pair to visualize}

\item{slot.name}{the slot name of object. Set is to be "netP" if input signaling is a pathway name; Set is to be "net" if input signaling is a ligand-receptor pair}

\item{pattern}{"outgoing" or "incoming"}

\item{zoom.to}{an integer specifying the zoom level for spatial visualization.
If \code{0}, no zooming is applied and the full field is shown.
If greater than 0, the spatial area is divided into \code{zoom.to + 1}
equal intervals along each axis}

\item{zoom.mode}{a logical value indicating how zooming is handled.
If \code{FALSE}, only the selected zoomed region is shown.
If \code{TRUE}, the selected region is highlighted while the rest of the
field remains visible with reduced transparency}

\item{anchor.x}{a numeric value between 0 and 1 specifying the relative horizontal position of the zoom anchor, used to select the zoomed region along the x-axis}

\item{anchor.y}{a numeric value between 0 and 1 specifying the relative vertical position of the zoom anchor, used to select the zoomed region along the y-axis}

\item{color.use}{defining the color for each cell group}

\item{point.size}{the size of spots}

\item{min.mag}{a numeric value specifying the minimum vector magnitude required for arrows to be displayed}

\item{alpha.image}{the transparency of individual spots}

\item{arrow.angle}{the angle of arrows}

\item{arrow.length}{the length of arrows}

\item{arrow.size}{the size of arrows}

\item{arrow.color}{the color of arrows}

\item{...}{other parameters passing to metR::geom_arrow}

\item{legend.size}{the size of legend}

\item{legend.text.size}{the text size on the legend}

\item{plot.title.name}{title name of the plot}

\item{plot.title.text.size}{the text size of the title}
}
\value{
ggplot2 object
}
\description{
netVisual_CommunField
}
