% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual_CommunFieldGrid}
\alias{netVisual_CommunFieldGrid}
\title{netVisual_CommunFieldGrid}
\usage{
netVisual_CommunFieldGrid(
  object,
  signaling,
  slot.name = "netP",
  pattern = c("incoming", "outgoing"),
  cellsize = NULL,
  what = "polygons",
  square = T,
  grid.resolution = NULL,
  grid.field.mean = c("median", "sum"),
  color.use = NULL,
  point.size = 2,
  image.alpha = 0.3,
  min.mag = 0,
  arrow.line.color = "black",
  arrow.line.alpha = 0.9,
  arrow.angle = 22.5,
  arrow.length = 0.5,
  ...,
  legend.size = 2,
  legend.text.size = 8,
  title.name = NULL
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{signaling}{a signaling pathway or ligand-receptor pair to visualize}

\item{slot.name}{the slot name of object. Set is to be "netP" if input signaling is a pathway name; Set is to be "net" if input signaling is a ligand-receptor pair}

\item{pattern}{"outgoing" or "incoming"}

\item{cellsize}{an integer specifying the spatial resolution (in pixels) used to construct the grid. If \code{NULL}, the minimum pairwise distance between spatial coordinates is used as the default cell size.}

\item{what}{a character string specifying the type of geometry returned by \code{sf::st_make_grid}. Default is "polygons"}

\item{square}{a logical value indicating whether the grid cells should be square}

\item{grid.resolution}{an integer specifying the scaling factor applied to \code{cellsize} when constructing the grid}

\item{grid.field.mean}{a character string specifying how vector fields within each grid cell are aggregated. Options include \code{"median"} or \code{"sum"}}

\item{color.use}{defining the color for each cell group}

\item{point.size}{the size of spots}

\item{image.alpha}{the transparency of individual spots}

\item{min.mag}{a numeric value specifying the minimum vector magnitude required for arrows to be displayed}

\item{arrow.line.color}{the color of arrows}

\item{arrow.line.alpha}{the transparency of arrows}

\item{arrow.angle}{the angle of arrows}

\item{arrow.length}{the length of arrows}

\item{...}{other parameters passing to metR::geom_arrow}

\item{legend.size}{the size of legend}

\item{legend.text.size}{the text size on the legend}

\item{title.name}{title name of the plot}
}
\value{
ggplot2 object
}
\description{
netVisual_CommunFieldGrid
}
