% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual_TopicComposition}
\alias{netVisual_TopicComposition}
\title{Stacked bar plot showing the proportion of cells across certrain cell groups}
\usage{
netVisual_TopicComposition(
  object,
  y.group,
  x.group = NULL,
  x.levels = NULL,
  y.levels = NULL,
  scale.rows = TRUE,
  cutoff.prop = 0.1,
  trim = 0.1,
  type = "heatmap",
  color.use = NULL,
  color.heatmap = "RdPu",
  n.colors = 8,
  dot.size = c(1, 6),
  angle.x = 45,
  xlabel = NULL,
  ylabel = NULL,
  title.name = NULL,
  legend.title = NULL,
  font.size = 8,
  font.size.title = 10,
  cluster.rows = TRUE,
  cluster.cols = FALSE,
  clustering_distance_rows = "euclidean",
  x.lab.rot = 45,
  row.show = NULL,
  col.show = NULL,
  remove.isolate = TRUE
)
}
\arguments{
\item{object}{SpatialCellChat object or seurat object}

\item{y.group}{grouping variable used to define rows of the topic composition matrix}

\item{x.group}{grouping variable used to define columns of the topic composition matrix}

\item{x.levels}{a character vector specifying the order of column groups}

\item{y.levels}{a character vector specifying the order of row groups}

\item{scale.rows}{whether to normalize proportions across rows.
When TRUE (default), values are scaled within each row to sum to 1.
When FALSE, values are scaled within each column.}

\item{cutoff.prop}{numeric threshold for filtering low-proportion groups}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each end of x before the mean is computed}

\item{type}{visualization type. Either "heatmap" (default) or "dot"}

\item{color.use}{defining the color for each column group}

\item{color.heatmap}{a character string or vector indicating the colormap option to use. It can be the avaibale color palette in brewer.pal() or viridis_pal()}

\item{n.colors}{Number of colors when setting colors.use to be a palette name from brewer.pal}

\item{dot.size}{numeric vector of length two specifying the minimum and maximum point sizes when `type = "dot"`}

\item{angle.x}{angle for x-axis text rotation}

\item{xlabel}{Name of x label}

\item{ylabel}{Name of y label}

\item{title.name}{Name of the main title}

\item{legend.title}{Name of legend}

\item{font.size}{font size in heatmap}

\item{font.size.title}{font size of the title}

\item{cluster.rows}{whether cluster rows}

\item{cluster.cols}{whether cluster columns}

\item{clustering_distance_rows}{distance metric used for row clustering in the heatmap}

\item{x.lab.rot}{rotation angle of x-axis tick labels}

\item{row.show, col.show}{a vector giving the index or the name of row or columns to show in the heatmap}

\item{remove.isolate}{whether remove the isolate nodes in the communication network}
}
\value{
ggplot2 object
}
\description{
Stacked bar plot showing the proportion of cells across certrain cell groups
}
