% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual_TopicSignaling}
\alias{netVisual_TopicSignaling}
\title{netVisual_TopicSignaling}
\usage{
netVisual_TopicSignaling(
  object,
  pattern = c("incoming", "outgoing"),
  slot.name = "net",
  topics = NULL,
  topic.key = "Topic_",
  nfeatures = 30,
  nTop = 10,
  feature_cat = c("LR-pair", "ligand", "signaling"),
  col = c("black", "black"),
  ncol = NULL,
  combine = T,
  ylabel = NULL,
  balanced = FALSE,
  text_nudge_x = -8,
  text_size = 3,
  point_size = 2,
  line_size = 1,
  species = NULL,
  db = NULL,
  simplify = TRUE
)
}
\arguments{
\item{object}{Seurat object or SpatialCellChat object}

\item{pattern}{"outgoing" or "incoming"}

\item{slot.name}{"net" or "netP".}

\item{topics}{NULL or Numeric. Determine which factors should be displayed.
By default, use `NULL` to display all factors' feature ranking plot.}

\item{topic.key}{character. By default, "Topic_"}

\item{nfeatures}{numeric. Determine how many top marker signalings in each factor to show}

\item{nTop}{numeric. Determine how many top marker signalings in each factor to show with a text annotation}

\item{feature_cat}{a character string specifying the feature category to summarize loadings, one of \code{"LR-pair"}, \code{"ligand"}, or \code{"signaling"}}

\item{col}{numeric of length 2. Colors of points to use, default is \code{c('neg'="#377EB8", 'pos'="#E41A1C")}}

\item{ncol}{number of columns in plot}

\item{combine}{combine plots into a single patchwork ggplot object. If FALSE, return a list of ggplot objects}

\item{ylabel}{name of y label}

\item{balanced}{return an equal number of signalings with + and - scores. If FALSE (default), returns the top signalings
ranked by the scores absolute values}

\item{text_nudge_x}{horizontal and vertical adjustments to nudge the starting position of each text label. The units for
nudge_x and nudge_y are the same as for the data units on the x-axis and y-axis}

\item{text_size}{numeric. Fontsize of text annotation}

\item{point_size}{numeric. Size of points}

\item{line_size}{numeric. Width of lines}

\item{species}{a character string specifying the species used to select the default CellChat database, either \code{"mouse"} or \code{"human"}}

\item{db}{an optional interaction database}

\item{simplify}{whether to combine results from all topics into a single data frame (TRUE) or return a list split by topic (FALSE)}
}
\value{
ggplot2 object
}
\description{
Ranking the signalings(LR pairs or pathways) and show the top marker signalings in each factor
}
