% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotFeatures}
\alias{plotFeatures}
\title{plotFeatures}
\usage{
plotFeatures(
  df_score,
  group.by = NULL,
  group.highlight = NULL,
  show.group = F,
  cells.highlight = NULL,
  method = c("normal", "3d", "2d"),
  plot.title = NULL,
  plot.ncol = 2,
  point.size = 3,
  highlight.alpha = 1,
  nohighlight.alpha = 0.65,
  show.colorscale = T,
  color.heatmap = "Blues",
  color.0 = NULL,
  color.NA = "#E5E5E5",
  n.colors = 8,
  direction = 1,
  color.use = NULL,
  legend.group = "score",
  legend.group.title = "score's value",
  scene.name = "scene1",
  colorbar.y = 0.5,
  three.dim.z = 0
)
}
\arguments{
\item{df_score}{a data frame with columns ordered as "x", "y", and "score".}

\item{group.by}{name of one or more metadata columns to group (color) cells by}

\item{group.highlight}{cell groups to be highlighted}

\item{show.group}{whether to display group annotations}

\item{cells.highlight}{cells/spots to be highlighted}

\item{method}{plotting mode, one of `"normal"` (ggplot2), `"2d`, or `"3d"` (Plotly)}

\item{plot.title}{title of the plot}

\item{plot.ncol}{the number of columns in the grid when `show.group = TRUE`}

\item{point.size}{the size of spots}

\item{highlight.alpha}{the opacity of highlighted cells/spots}

\item{nohighlight.alpha}{the opacity of non-highlighted cells/spots}

\item{show.colorscale}{whether to display the colorbar}

\item{color.heatmap}{A character string or vector indicating the colormap option to use. It can be the avaibale color palette in brewer.pal() or viridis_pal()}

\item{color.0}{color used for 0 in the colormap}

\item{color.NA}{color used to represent missing values (NA) in the plot}

\item{n.colors}{number of basic colors to generate from color palette}

\item{direction}{sets the order of colors in the scale. If 1, the default colors are used. If -1, the order of colors is reversed.}

\item{color.use}{defining the color for each cell group}

\item{legend.group}{the parameter `legendgroup` in `plotly::plot_ly` and `plotly::add_trace`. Name of the legend group for traces, used to combine multiple traces under the same legend.}

\item{legend.group.title}{the parameter `legendgrouptitle` in `plotly::plot_ly` and `plotly::add_trace`. The title text displayed for the legend group.}

\item{scene.name}{name of the Plotly 3D scene}

\item{colorbar.y}{the relative position of the color bar along the y-axis}

\item{three.dim.z}{the fixed z-axis coordinate in 3D mode}
}
\value{
ggplot2 object or plotly object
}
\description{
plotFeatures
}
