% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotly_spatialDimPlot}
\alias{plotly_spatialDimPlot}
\title{interactive spatialDimPlot}
\usage{
plotly_spatialDimPlot(
  object,
  group.by = NULL,
  group.highlight = NULL,
  cells.highlight = NULL,
  sources.use = NULL,
  targets.use = NULL,
  idents.use = NULL,
  method = c("2d", "3d"),
  color.use = NULL,
  nohighlight.alpha = 0.65,
  title.name = NULL,
  scene.name = "scene",
  point.size = 2,
  show.legend = T,
  legend.group.title = "labels",
  three.dim.z = 0
)
}
\arguments{
\item{object}{a SpatialCellChat object}

\item{group.by}{name of one or more metadata columns to group (color) cells by}

\item{group.highlight}{cell groups to be highlighted}

\item{cells.highlight}{cells/spots to be highlighted}

\item{sources.use}{a vector giving the index or the name of source cell groups of interest}

\item{targets.use}{a vector giving the index or the name of target cell groups of interest}

\item{idents.use}{a vector giving the index or the name of cell groups of interest}

\item{method}{plotting mode, one of `"2d`, or `"3d"` (Plotly)}

\item{color.use}{defining the color for each cell group}

\item{nohighlight.alpha}{the opacity of non-highlighted cells/spots}

\item{title.name}{set the title of this plot}

\item{scene.name}{name of the Plotly 3D scene}

\item{point.size}{the size of spots}

\item{show.legend}{whether show each figure legend}

\item{legend.group.title}{the parameter `legendgrouptitle` in `plotly::plot_ly` and `plotly::add_trace`. The title text displayed for the legend group.}

\item{three.dim.z}{the fixed z-axis coordinate in 3D mode}

\item{legend.group}{the parameter `legendgroup` in `plotly::plot_ly` and `plotly::add_trace`. Name of the legend group for traces, used to combine multiple traces under the same legend.}
}
\value{
a plotly plot
}
\description{
interactive spatialDimPlot
}
