% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotly_spatialLRpairPlot}
\alias{plotly_spatialLRpairPlot}
\title{plotly_spatialLRpairPlot}
\usage{
plotly_spatialLRpairPlot(
  object,
  features = NULL,
  pairLR.use = NULL,
  cells.highlight = NULL,
  enriched.only = F,
  thresh = 0.05,
  do.group = F,
  colormap = "Spectral",
  reversescale = T,
  point.size = 4,
  nohighlight.alpha = 0.3,
  do.plot = T
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{features}{a char vector containing features to visualize. `features` can be genes or column names of `object@meta`.}

\item{pairLR.use}{a char vector or a data frame consisting of one column named "interaction_name", defining the L-R pairs of interest}

\item{cells.highlight}{cells/spots to be highlighted. It can be cell IDs (character vector) or indices (numeric vector)}

\item{enriched.only}{whether only return the identified enriched signaling genes in the database. Default = TRUE, returning the significantly enriched signaling interactions}

\item{thresh}{threshold of the p-value for determining significant interaction}

\item{do.group}{set `do.group = TRUE` when computing scores of each cell group; set `do.group = FALSE` when computing scores of each individual cell}

\item{colormap}{RColorbrewer palette to use (check available palette using RColorBrewer::display.brewer.all()). default will use customed color palette}

\item{reversescale}{whether to reverse the direction of the color scale}

\item{point.size}{the size of spots}

\item{nohighlight.alpha}{the opacity of non-highlighted cells/spots}

\item{do.plot}{whether to generate and return plots}
}
\value{
a list of ggplot2/plotly objects (`do.plot = TRUE`) or the number of features (`do.plot = FALSE`)
}
\description{
plotly_spatialLRpairPlot
}
