% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotly_spatialLRpairPlot_shiny}
\alias{plotly_spatialLRpairPlot_shiny}
\title{plotly_spatialLRpairPlot_shiny}
\usage{
plotly_spatialLRpairPlot_shiny(
  object,
  features = NULL,
  pairLR.use = NULL,
  point.size.dimplot = 2,
  point.size.LRplot = 6,
  nohighlight.alpha = 0.45,
  port = 10333
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{features}{a char vector containing features to visualize. `features` can be genes or column names of `object@meta`.}

\item{pairLR.use}{a char vector or a data frame consisting of one column named "interaction_name", defining the L-R pairs of interest}

\item{point.size.dimplot}{point size used in the spatialDimPlot}

\item{point.size.LRplot}{point size used in the spatialLRpairPlot}

\item{nohighlight.alpha}{the opacity of non-highlighted cells/spots}

\item{port}{port number used to launch the Shiny application}
}
\value{
A Shiny application object
}
\description{
plotly_spatialLRpairPlot_shiny
}
