% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotStatistics_Gi}
\alias{plotStatistics_Gi}
\title{plotStatistics_Gi}
\usage{
plotStatistics_Gi(
  df_score,
  group.by,
  group.highlight,
  n,
  method = "normal",
  gstar = FALSE,
  p.value = NULL,
  plot.title = NULL,
  point.size = 2,
  nohighlight.alpha = 0.65,
  show.colorscale = T,
  legend.group = "Gi",
  legend.group.title = "Gi's value",
  scene.name = "scene1",
  colorbar.y = 0.5,
  three.dim.z = 0
)
}
\arguments{
\item{df_score}{a data frame with columns ordered as "x", "y", and "score"}

\item{group.by}{name of one or more metadata columns to group (color) cells by}

\item{group.highlight}{cell groups to be highlighted}

\item{n}{the number of nearest neighbours for each spot, used to compute the Gi statistic}

\item{method}{plotting mode. one of "scoring", "normal", "3d", "density" or "contour"}

\item{gstar}{logical; If true, compute the Gi* statistic, else compute the Gi statistic}

\item{p.value}{significance threshold for p-values}

\item{plot.title}{the title of the plot}

\item{point.size}{the size of spots}

\item{nohighlight.alpha}{the opacity of non-highlighted cells/spots}

\item{show.colorscale}{whether to display the colorbar}

\item{legend.group}{the parameter `legendgroup` in `plotly::plot_ly` and `plotly::add_trace`. Name of the legend group for traces, used to combine multiple traces under the same legend}

\item{legend.group.title}{the parameter `legendgrouptitle` in `plotly::plot_ly` and `plotly::add_trace`. The title text displayed for the legend group}

\item{scene.name}{name of the Plotly 3D scene}

\item{colorbar.y}{the relative position of the color bar along the y-axis}

\item{three.dim.z}{the fixed z-axis coordinate in 3D mode}
}
\value{
a ggplot2 object, a plotly object or a data frame
}
\description{
plotStatistics_Gi
}
