% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotStatistics_Lee}
\alias{plotStatistics_Lee}
\title{plotStatistics_Lee}
\usage{
plotStatistics_Lee(
  x.labels,
  y.labels,
  coords = NULL,
  lw = NULL,
  x.levels = NULL,
  y.levels = NULL,
  cutoff = TRUE,
  cutoff.Lee = 0,
  type = "heatmap",
  color.use = NULL,
  color.heatmap = "Reds",
  reverse.colorscale = F,
  n.colors = 6,
  dot.size = c(1, 6),
  angle.x = 45,
  xlabel = NULL,
  ylabel = NULL,
  title.name = NULL,
  legend.title = NULL,
  font.size = 10,
  font.size.title = 10,
  x.lab.rot = 45,
  row.show = NULL,
  col.show = NULL
)
}
\arguments{
\item{x.labels}{labels defining the categories for the x variables. Can be a factor/vector, a numeric vector, or a matrix}

\item{y.labels}{labels defining the categories for the y variables. Can be a factor/vector, a numeric vector, or a matrix}

\item{coords}{a two-column matrix or data frame specifying spatial coordinates}

\item{lw}{spatial distance weight matrix}

\item{x.levels}{a character vector specifying the order of x categories}

\item{y.levels}{a character vector specifying the order of y categories}

\item{cutoff}{whether to apply a threshold to Lee's statistics}

\item{cutoff.Lee}{numeric threshold for Lee's statistic. Values less than or equal to this cutoff are removed when `cutoff = TRUE`}

\item{type}{the type of the output, one of "data", "dot" or "heatmap"}

\item{color.use}{defining the color for each group}

\item{color.heatmap}{A character string or vector indicating the colormap option to use. It can be the avaibale color palette in brewer.pal() or viridis_pal() (e.g., "Spectral","viridis")}

\item{reverse.colorscale}{whether to reverse the colorbar}

\item{n.colors}{number of basic colors to generate from color palette}

\item{dot.size}{a range defining the size of the symbol}

\item{angle.x}{angle for x-axis text rotation}

\item{xlabel}{label of x-axis}

\item{ylabel}{label of y-axis}

\item{title.name}{title name}

\item{legend.title}{name of legend}

\item{font.size}{font size in the plot}

\item{font.size.title}{font size of the title}

\item{x.lab.rot}{rotation angle of x-axis tick labels}

\item{row.show, col.show}{a vector giving the index or the name of row or columns to show in the heatmap}
}
\value{
a data frame, a ggplot2 object or a Heatmap-class object
}
\description{
plotStatistics_Lee
}
