% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{relabelSpatialCellChat}
\alias{relabelSpatialCellChat}
\title{relabelSpatialCellChat}
\usage{
relabelSpatialCellChat(
  object,
  labelSet,
  group.by = NULL,
  nboot = 100,
  min.cells.sr = 10,
  min.percent = 0.1,
  do.permutation = T,
  min.cells = 10,
  thresh = 0.05
)
}
\arguments{
\item{object}{SpatialCellChat object.}

\item{labelSet}{Named character vector. Make sure the names are cell group labels in
present `object@idents`, and values are new cell group labels respectively.
For example,`labelSet = c("Group1"="NewGroup1","Group2"="NewGroup1","Group3"="NewGroup1")`.}

\item{group.by}{Set a column name in `object@meta` if not use `object@idents` as cell group labels
to relabel.}

\item{nboot}{Integer. See the same parameter in \link{computeAvgCommunProb} or \link{computeAvgCommunProb_Visium}.}

\item{min.cells.sr}{Integer. See the same parameter in \link{computeAvgCommunProb}.}

\item{min.percent}{Numeric. See the same parameter in \link{computeAvgCommunProb}.}

\item{do.permutation}{Boolean. Integer. See the same parameter in \link{computeAvgCommunProb} or \link{computeAvgCommunProb_Visium}.}

\item{min.cells}{Integer. See the same parameter in \link{filterCommunication}.}

\item{thresh}{Numeric. See the same parameter in \link{computeCommunProbPathway}.}
}
\value{
SpatialCellChat object.
}
\description{
Merge some cell groups into a new cell group, and recalculate cell-cell communication
probability/strength at the cell-group level.
}
