% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RunSeurat}
\alias{RunSeurat}
\title{RunSeurat}
\usage{
RunSeurat(
  Obejct,
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  selection.nfeatures = 2000,
  selection.method = "vst",
  do.center = T,
  do.scale = T,
  PCA.npcs = 50,
  k_Neighbors = 20,
  dims_Neighbors = 1:40,
  dims_UMAP = 1:10,
  seed.use = 666L,
  verbose = TRUE
)
}
\arguments{
\item{Obejct}{A seurat object}

\item{normalization.method}{See details of `normalization.method` in \code{\link[Seurat]{NormalizeData.Seurat}}.}

\item{scale.factor}{See details of `scale.factor` in \code{\link[Seurat]{NormalizeData.Seurat}}.}

\item{selection.nfeatures}{See details of `nfeatures` in \code{\link[Seurat]{FindVariableFeatures}}.}

\item{selection.method}{See details of `selection.method` in \code{\link[Seurat]{FindVariableFeatures}}.}

\item{do.center}{See details of `do.center` in \code{\link[Seurat]{ScaleData}}.}

\item{do.scale}{See details of `do.scale` in \code{\link[Seurat]{ScaleData}}.}

\item{PCA.npcs}{See details of `npcs`  in \code{\link[Seurat]{RunPCA}}.}

\item{k_Neighbors}{See details of `k.param` in \code{\link[Seurat]{FindNeighbors}}.}

\item{dims_Neighbors}{See details of `dims` in \code{\link[Seurat]{FindNeighbors}}.}

\item{dims_UMAP}{See details of `dims` in \code{\link[Seurat]{RunUMAP}}.}

\item{seed.use}{Seed used for random number generation in PCA & UMAP.}

\item{verbose}{Controls verbosity.}
}
\value{
A seurat object
}
\description{
RunSeurat
}
