% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{savePlotWithCow}
\alias{savePlotWithCow}
\title{savePlotWithCow}
\usage{
savePlotWithCow(
  filename,
  plot = NULL,
  ncol = 1,
  nrow = 1,
  base_height = 3.71,
  base_asp = 1.618,
  base_width = NULL,
  ...,
  cols,
  rows,
  base_aspect_ratio,
  width,
  height
)
}
\arguments{
\item{filename}{Name of the plot file to generate.}

\item{plot}{Plot to save.}

\item{ncol}{Number of subplot columns.}

\item{nrow}{Number of subplot rows.}

\item{base_height}{The height (in inches) of the plot or of one sub-plot if \code{nrow}
or \code{ncol} > 1. Default is 3.71.}

\item{base_asp}{The aspect ratio (width/height) of the plot or of one sub-plot if \code{nrow}
or \code{ncol} > 1. This argument is used if \code{base_width = NULL} or if \code{base_height = NULL};
if both width and height are provided then the aspect ratio is ignored.
The default is 1.618 (the golden ratio), which works well for figures with a legend.}

\item{base_width}{The width (in inches) of the plot or of one sub-plot if \code{nrow}
or \code{ncol} > 1. Default is \code{NULL}, which means that the width is calculated from
\code{base_height} and \code{base_aspect_ratio}.}

\item{...}{Other arguments to be handed to \code{\link[cowplot:ggsave2]{ggsave2()}}.}

\item{cols}{Deprecated. Use \code{ncol}.}

\item{rows}{Deprecated. Use \code{nrow}.}

\item{base_aspect_ratio}{Deprecated. Use \code{base_asp}.}

\item{width}{Deprecated. Don't use.}

\item{height}{Deprecated. Don't use.}
}
\description{
Save a plot directly from `Plots` window in Rstudio with the help of \code{cowplot}.
See details in \code{\link[cowplot]{save_plot}}.
}
