% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{scMatrixTruncation}
\alias{scMatrixTruncation}
\title{scMatrixTruncation}
\usage{
scMatrixTruncation(Mat, cutoff = NULL, remain.cutoff.v = T, repr = c("C", "T"))
}
\arguments{
\item{Mat}{dgCMatrix or dgTMatrix}

\item{cutoff}{Numeric. Cut off Expression Matrix's values to make
Mat' values >= cutoff.}

\item{remain.cutoff.v}{Boolean. Use `>=` or `>`.}

\item{repr}{Character string. One of "C", "T", specifying the representation of the sparse matrix result.}
}
\value{
dgCMatrix or dgTMatrix
}
\description{
Single cell data is usually saved as "dgCMatrix" or "dgTMatrix" in R,
use this function to cut off Expression Matrix's values to make
Mat'v >= cutoff.
}
