% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setEnvironment}
\alias{setEnvironment}
\title{Set the computation environment}
\usage{
setEnvironment(
  workers,
  future.globals.maxSize = 10000 * 1024^2,
  gc = T,
  conda_env = NULL
)
}
\arguments{
\item{workers}{A integer greater than 0. See details in \code{\link[future]{multisession}}.
When workers = 1, CellChat will run in sequential mode;
Otherwise, CellChat will run in parallel mode with a respective number of future parallel clusters.}

\item{future.globals.maxSize}{Run \code{?future::future.globals.maxSize} to see details.}

\item{gc}{Boolean. Whether the garbage collector run or not. It's an argument passed to Future().
See details in \code{\link[future]{Future}}}

\item{conda_env}{We use some functions coded with Python in CellChat. Users
may need to set Python Lib Path here or otherwhere explicitly. For example, setEnvironment(workers = 2L,conda_env = "path/to/.conda/envs/env_name/")
See details in \code{\link[reticulate]{use_condaenv}} or use `reticulate`'s other relevant functions to get a Python support.}
}
\description{
Set the computation environment
}
