% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setpwd}
\alias{setpwd}
\title{Set working directory to the path of currently opened file.}
\usage{
setpwd()
}
\description{
Set working directory to the path of currently opened file (usually an R script). You can use this function in both .R/.Rmd files and R Console. RStudio (version >= 1.2) is required for running this function.
}
