% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCellChat_class.R
\docType{methods}
\name{show,SpatialCellChat-method}
\alias{show,SpatialCellChat-method}
\title{show method for SpatialCellChat}
\usage{
\S4method{show}{SpatialCellChat}(object)
}
\arguments{
\item{object}{object}

\item{SpatialCellChat}{object}

\item{show}{show the object}
}
\description{
show method for SpatialCellChat
}
