% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatialCCCDistPlot}
\alias{spatialCCCDistPlot}
\title{Spatial Cell-Cell Communication Distance Plot}
\usage{
spatialCCCDistPlot(
  object,
  signaling.type = c("both", "Secreted", "Contact"),
  interaction.range = 250,
  contact.range = 10,
  tol = NULL,
  density.alpha = 0.5
)
}
\arguments{
\item{object}{SpatialCellChat object.}

\item{signaling.type}{Character. One of c("both","Secreted","Contact").}

\item{interaction.range}{Numeric. By default 250um. The maximum interaction/diffusion length of ligands (Unit: microns).
This hard threshold is used to filter out the connections between spatially distant individual cells.}

\item{contact.range}{Numeric. By default 10um. The interaction range (Unit: microns) to restrict the contact-dependent signaling.
For spatial transcriptomics in a single-cell resolution, `contact.range` is approximately equal to the estimated
cell diameter (i.e., the cell center-to-center distance), which means that contact-dependent and juxtacrine
signaling can only happens when the two cells are contact to each other.}

\item{tol}{NULL or Numeric. set a distance tolerance when computing cell-cell distances and contact adjacent matrix.
By default `tol = NULL` means distance tolerance equals to spot.size/2 or cell.diameter/2.}

\item{density.alpha}{Numeric. Control the transparency of the plot.}
}
\value{
ggplot
}
\description{
SpatialCellChat uses cell-cell distances as communication constraint. Use this
function to visualize reasonable cell-cell communication distances in the
spatial transcriptomics data for both "Secreted" signalings and "Contact" signalings.
}
