% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatialDimPlot}
\alias{spatialDimPlot}
\title{Visualize spatial cell groups such as signaling sources and targets}
\usage{
spatialDimPlot(
  object,
  color.use = NULL,
  group.by = NULL,
  sources.use = NULL,
  targets.use = NULL,
  idents.use = NULL,
  proportion = NULL,
  radius = NULL,
  radius.size = 0.55,
  alpha = 1,
  shape.by = 16,
  title.name = NULL,
  point.size = 2.4,
  legend.size = 3,
  legend.text.size = 8,
  legend.position = "right",
  legend.spacing = c(0, -8),
  ncol = 1,
  byrow = FALSE
)
}
\arguments{
\item{object}{cellchat object}

\item{color.use}{defining the color for each cell group}

\item{group.by}{Name of one metadata columns to group (color) cells. Default is the defined cell groups in CellChat object}

\item{sources.use}{a vector giving the index or the name of source cell groups}

\item{targets.use}{a vector giving the index or the name of target cell groups}

\item{idents.use}{a vector giving the index or the name of cell groups of interest}

\item{proportion}{data frame of cell type proportion (rows are cells and columns are cell types)}

\item{radius}{dumeric value about the radius of each pie chart, if NULL, we will calculate it inside the function}

\item{radius.size}{numeric scaling factor for the pie chart radius}

\item{alpha}{the transparency of individual spot}

\item{shape.by}{the shape of individual spot}

\item{title.name}{title name}

\item{point.size}{the size of spots}

\item{legend.size}{the size of legend}

\item{legend.text.size}{the text size on the legend}

\item{legend.position, ncol, byrow}{parameters for configurating the plot}

\item{legend.spacing}{a two-elements vector respectively specifying legend.key.spacing.x and legend.key.spacing.y for spacing apart legend key-label pairs}
}
\description{
This function takes a CellChat object as input, and then plot cell groups of interest.
}
