% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatialGiPlot}
\alias{spatialGiPlot}
\title{spatialGiPlot}
\usage{
spatialGiPlot(
  object,
  slot.name = "netP",
  signaling.name = NULL,
  do.binary = T,
  group.highlight = NULL,
  measure = c("indeg", "outdeg"),
  measure.name = c("incoming", "outgoing"),
  plot.title = NULL,
  n = 12,
  method = "normal",
  gstar = FALSE,
  p.value = NULL,
  three.dim.z = 0,
  point.size = 2,
  nohighlight.alpha = 0.65,
  show.colorscale = T,
  return.object = T,
  show.plot = T
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{slot.name}{the slot name of object that is used for analysis}

\item{signaling.name}{alternative signaling pathway name to show on the plot}

\item{do.binary}{whether to plot the scores in a binary mode}

\item{group.highlight}{cell groups to be highlighted}

\item{measure}{"indeg" or "outdeg". "indeg": in-degree of the communication network. "outdeg": out-degree of the communication network}

\item{measure.name}{the measure names to show, which should correspond to the defined `measure`}

\item{plot.title}{title of the plot}

\item{n}{the number of nearest neighbours for each spot, used to compute the Gi statistic}

\item{method}{plotting mode. one of "scoring", "normal", "3d", "density" or "contour"}

\item{gstar}{logical; If true, compute the Gi* statistic, else compute the Gi statistic}

\item{p.value}{significance threshold for p-values}

\item{three.dim.z}{the fixed z-axis coordinate in 3D mode}

\item{point.size}{the size of spots}

\item{nohighlight.alpha}{the opacity of non-highlighted cells/spots}

\item{show.colorscale}{whether to display the colorbar}

\item{return.object}{return either SpatialCellChat object or ggplot2 object}

\item{show.plot}{if `return.object = TRUE`, whether to display the plot}
}
\value{
SpatialCellChat object, a ggplot2 object or a plotly object
}
\description{
spatialGiPlot
}
