% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatialLeePlot}
\alias{spatialLeePlot}
\title{spatialLeePlot}
\usage{
spatialLeePlot(
  object,
  group.by2,
  group.by1 = NULL,
  x.levels = NULL,
  y.levels = NULL,
  lw.type = c("Itself", "contact.range", "interaction.range"),
  lw = NULL,
  interaction.range = 250,
  contact.range = 10,
  cutoff = TRUE,
  cutoff.Lee = 0,
  type = "heatmap",
  color.use = NULL,
  color.heatmap = "Reds",
  reverse.colorscale = F,
  n.colors = 6,
  dot.size = c(1, 6),
  angle.x = 45,
  xlabel = NULL,
  ylabel = NULL,
  title.name = NULL,
  legend.title = NULL,
  font.size = 10,
  font.size.title = 10,
  x.lab.rot = 45,
  row.show = NULL,
  col.show = NULL
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{group.by2}{a character string specifying a column name in `object@meta` used to define the grouping variable for x.
Alternatively, a vector of labels can be provided directly}

\item{group.by1}{a character string specifying a column name in `object@meta` used to define the grouping variable for y.
If NULL, `object@idents` is used by default. Alternatively, a vector of labels can be provided directly}

\item{x.levels}{a character vector specifying the order of x categories}

\item{y.levels}{a character vector specifying the order of y categories}

\item{lw.type}{select the type of distance weight matrix based on the communication mode, one of "Itself","contact.range", or "interaction.range"}

\item{lw}{spatial distance weight matrix. If `lw = NULL`, it will be automatically computed based on `lw.type`. Default is `NULL`}

\item{interaction.range}{numeric. By default 250um. The maximum interaction/diffusion length of ligands (Unit: microns)
This hard threshold is used to filter out the connections between spatially distant individual cells}

\item{contact.range}{numeric. By default 10um. The interaction range (Unit: microns) to restrict the contact-dependent signaling.
For spatial transcriptomics in a single-cell resolution, `contact.range` is approximately equal to the estimated cell diameter (i.e., the cell center-to-center distance), which means that contact-dependent and juxtacrine signaling can only happens when the two cells are contact to each other}

\item{cutoff}{whether to apply a threshold to Lee's statistics}

\item{cutoff.Lee}{numeric threshold for Lee's statistic. Values less than or equal to this cutoff are removed when `cutoff = TRUE`}

\item{type}{the type of the output, one of "data", "dot" or "heatmap"}

\item{color.use}{defining the color for each group}

\item{color.heatmap}{A character string or vector indicating the colormap option to use. It can be the avaibale color palette in brewer.pal() or viridis_pal() (e.g., "Spectral","viridis")}

\item{reverse.colorscale}{whether to reverse the colorbar}

\item{n.colors}{number of basic colors to generate from color palette}

\item{dot.size}{a range defining the size of the symbol}

\item{angle.x}{angle for x-axis text rotation}

\item{xlabel}{label of x-axis}

\item{ylabel}{label of y-axis}

\item{title.name}{title name}

\item{legend.title}{name of legend}

\item{font.size}{font size in the plot}

\item{font.size.title}{font size of the title}

\item{x.lab.rot}{rotation angle of x-axis tick labels}

\item{row.show, col.show}{a vector giving the index or the name of row or columns to show in the heatmap}
}
\value{
a ggplot2 object
}
\description{
spatialLeePlot
}
