% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatialTopicPlot}
\alias{spatialTopicPlot}
\title{SpatialTopicPlot}
\usage{
spatialTopicPlot(
  object,
  topics = NULL,
  slot.name = "net",
  pattern = c("incoming", "outgoing"),
  topics.key = "Topic_",
  color.heatmap = "Reds",
  n.colors = 8,
  direction = 1,
  cutoff = NULL,
  color.use = NULL,
  alpha = 1,
  point.size = 1,
  legend.size = 3,
  legend.text.size = 8,
  shape.by = 16,
  ncol = NULL,
  combine = T,
  show.legend = TRUE,
  show.legend.combined = FALSE
)
}
\arguments{
\item{object}{SpatialCellChat object}

\item{topics}{a vector of numeric indices or character names specifying which spatial topics to visualize. If \code{NULL}, all topics available are plotted}

\item{slot.name}{the slot name of object that is used for topic analysis}

\item{pattern}{"outgoing" or "incoming"}

\item{topics.key}{a character string used as a prefix to generate topic names}

\item{color.heatmap}{a character string or vector indicating the colormap option to use. It can be the avaibale color palette in brewer.pal() or viridis_pal()}

\item{n.colors}{number of basic colors to generate from color palette}

\item{direction}{sets the order of colors in the scale. If 1, the default colors are used. If -1, the order of colors is reversed}

\item{cutoff}{a numeric value used to filter low loading values}

\item{color.use}{defining the color for each cell group}

\item{alpha}{the transparency of individual spot}

\item{point.size}{the size of spots}

\item{legend.size}{the size of legend}

\item{legend.text.size}{the text size on the legend}

\item{shape.by}{the shape of individual spot}

\item{ncol}{number of columns if plotting multiple plots}

\item{combine}{whether outputting an combined plot using patchwork}

\item{show.legend}{whether show each figure legend}

\item{show.legend.combined}{whether show the figure legend for the last plot}
}
\value{
a ggplot2 object
}
\description{
SpatialTopicPlot
}
