% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatialVisual_scoring}
\alias{spatialVisual_scoring}
\title{Spatially visualizing any computed scores}
\usage{
spatialVisual_scoring(
  object,
  signaling,
  method = "centrality",
  measure = c("outdeg", "indeg"),
  measure.name = c("outgoing", "incoming"),
  merge = FALSE,
  slot.name = "netP",
  do.group = FALSE,
  color.heatmap = "BuGn",
  n.colors = 9,
  direction = 1,
  alpha = 1,
  do.binary = FALSE,
  color.use = NULL,
  point.size = 0.8,
  legend.size = 3,
  legend.text.size = 8,
  legend.spacing = c(0, -8),
  shape.by = 16,
  ncol = NULL,
  combined = TRUE,
  show.legend = TRUE,
  show.legend.combined = FALSE
)
}
\arguments{
\item{object}{cellchat object}

\item{signaling}{a signaling pathway or ligand-receptor pair to visualize}

\item{method}{the method for scoring}

\item{measure}{measurements to show. When method = "centrality", centality measurements can be c("outdeg_unweighted", "indeg_unweighted","outdeg","indeg","hub","authority","eigen","page_rank","betweenness","flowbet","info")}

\item{measure.name}{the measure names to show, which should correspond to the defined `measure`}

\item{merge}{whether to merge two scoring into one single plot}

\item{slot.name}{the slot name of object that is used to compute scores of signaling networks}

\item{do.group}{set `do.group = TRUE` when computing scores of each cell group; set `do.group = FALSE` when computing scores of each individual cell}

\item{color.heatmap}{A character string or vector indicating the colormap option to use. It can be the avaibale color palette in viridis_pal() or brewer.pal()}

\item{n.colors, direction}{n.colors: number of basic colors to generate from color palette; direction: Sets the order of colors in the scale. If 1, the default colors are used. If -1, the order of colors is reversed.}

\item{alpha}{the transparency of individual spot}

\item{do.binary}{whether to plot the scores in a binary mode}

\item{color.use}{colors to use when `do.binary = TRUE`}

\item{point.size}{the size of cell slot}

\item{legend.size}{the size of legend}

\item{legend.text.size}{the text size on the legend}

\item{legend.spacing}{a two-elements vector respectively specifying legend.key.spacing.x and legend.key.spacing.y for spacing apart legend key-label pairs}

\item{shape.by}{the shape of individual spot}

\item{ncol}{number of columns when plotting multiple plots}

\item{combined}{whether outputting an combined plot using patchwork}

\item{show.legend}{whether show each figure legend}

\item{show.legend.combined}{whether show the figure legend for the last plot}
}
\description{
Spatially visualizing any computed scores
}
