% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{suggestK1}
\alias{suggestK1}
\title{suggestK1}
\usage{
suggestK1(
  Mat,
  k.test = seq(5, 50, 5),
  reps = 10,
  p = 0.3,
  tol = 1e-06,
  maxit = 100L,
  L1 = c(0, 0),
  seed.use = 666L
)
}
\arguments{
\item{Mat}{any `Matrix` object}

\item{k.test}{integer vector. Choose several ks for NMF cross validation}

\item{reps}{integer. How many replications to use for each k in cross validation}

\item{p}{numeric. Proportion(0<p<1) of the Mat's elements used for NMF cross validation}

\item{tol}{numeric. Stopping criteria. See the same parameter in \code{\link[RcppML]{nmf}}}

\item{maxit}{integer. Stopping criteria. See the same parameter in \code{\link[RcppML]{nmf}}}

\item{L1}{L1/LASSO penalties between 0 and 1, array of length two for c(w, h)}

\item{seed.use}{integer. Random seed for model initialization}
}
\value{
a list containing a data frame of cross-validation statistics across tested NMF ranks and a diagnostic plot to guide the selection of the optimal NMF dimensionality
}
\description{
Cross validate for RcppML::nmf.
}
