% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{themeFeaturePlot}
\alias{themeFeaturePlot}
\title{themeFeaturePlot}
\usage{
themeFeaturePlot(legend.text.size = 8)
}
\arguments{
\item{legend.text.size}{the text size on the legend}
}
\value{
a ggplot2 theme object
}
\description{
Defines a reusable ggplot2 theme for feature plots
}
