% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCellChat_class.R
\name{updateSpatialCellChat}
\alias{updateSpatialCellChat}
\title{Update a single SpatialCellChat object}
\usage{
updateSpatialCellChat(object)
}
\arguments{
\item{object}{SpatialCellChat object}
}
\value{
a updated SpatialCellChat object
}
\description{
Update a single previously calculated SpatialCellChat object (version < 1.6.0)
}
\details{
version < 0.5.0: `object@var.features` is now `object@var.features$features`; `object@net$sum` is now `object@net$weight` if `aggregateNet` has been run.

version 1.6.0: a `object@images` slot is added and `datatype` is added in `object@options$datatype`

version 2.0.0: `images$scale.factors` is changed to `images$spatial.factors` for spatial transcriptomics data analysis.
}
