% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizDimLoadings}
\alias{vizDimLoadings}
\title{vizDimLoadings}
\usage{
vizDimLoadings(
  object,
  dims = 1:5,
  nfeatures = 30,
  nTop = 10,
  feature_cat = c("LR-pair", "ligand", "signaling"),
  col = c(neg = "#377EB8", pos = "#E41A1C"),
  reduction = "pca",
  projected = FALSE,
  balanced = FALSE,
  species = c("mouse", "human"),
  db = NULL,
  ncol = NULL,
  combine = TRUE,
  ylabel = "Signaling score",
  text_nudge_x = -8,
  text_size = 3,
  point_size = 2,
  line_size = 1
)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{an integer vector specifying which dimensions to analyze}

\item{nfeatures}{an integer specifying the number of top features to extract from each dimension}

\item{nTop}{an integer specifying the number of top-ranked features to show with a text annotation}

\item{feature_cat}{a character string specifying the feature category to summarize loadings, one of \code{"LR-pair"}, \code{"ligand"}, or \code{"signaling"}}

\item{col}{numeric of length 2. Colors of points to use, default is \code{c('neg'="#377EB8", 'pos'="#E41A1C")}}

\item{reduction}{a character string specifying the dimensional reduction method stored in the Seurat object}

\item{projected}{whether to use the projected feature loadings}

\item{balanced}{whether to balance positive and negative features}

\item{species}{a character string specifying the species used to select the default CellChat database, either \code{"mouse"} or \code{"human"}}

\item{db}{an optional interaction database}

\item{ncol}{number of columns to show in the plot}

\item{combine}{whether outputting an combined plot using patchwork}

\item{ylabel}{name of y label}

\item{text_nudge_x}{horizontal and vertical adjustments to nudge the starting position of each text label. The units for
nudge_x and nudge_y are the same as for the data units on the x-axis and y-axis}

\item{text_size}{fontsize of text annotation}

\item{point_size}{numeric. Size of points}

\item{line_size}{numeric. Width of lines}
}
\value{
ggplot2 object
}
\description{
vizDimLoadings
}
