% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{preProcessing}
\alias{preProcessing}
\title{preProcessing}
\usage{
preProcessing(object, slot.name = c("data.signaling", "data"))
}
\arguments{
\item{object}{Matrix or SpatialCellChat object.}

\item{slot.name}{Character. By default "data.signaling". Use "data.signaling" or "data" when `object` is a SpatialCellChat object.}
}
\value{
Matrix or SpatialCellChat object.
}
\description{
Use this function to pre-process the data matrix input before running \link{createSpatialCellChat},
or pre-process the SpatialCellChat object following \link{subsetData} function.
}
